Update sync args

+++ gh-gl-sync/SpackCIBridge.py
@@ -360,7 +360,8 @@ class SpackCIBridge(object):
                 subprocess.run(push_args, check=True)
 
             # Clean up per-PR dedicated mirrors for any closed PRs
-            self.delete_pr_mirrors(closed_refspecs)
+            if "AWS_ACCESS_KEY_ID" in os.environ and "AWS_SECRET_ACCESS_KEY" in os.environ:
+                self.delete_pr_mirrors(closed_refspecs)
 
             # Post pipeline status to GitHub for each open PR.
             pipeline_api_template = self.get_pipeline_api_template(args.gitlab_host, args.gitlab_project)
@@ -383,12 +384,6 @@ if __name__ == "__main__":
 
     if "GITHUB_TOKEN" not in os.environ:
         raise Exception("GITHUB_TOKEN environment is not set")
-
-    if "AWS_ACCESS_KEY_ID" not in os.environ:
-        raise Exception("AWS_ACCESS_KEY_ID environment is not set")
-
-    if "AWS_SECRET_ACCESS_KEY" not in os.environ:
-        raise Exception("AWS_SECRET_ACCESS_KEY environment is not set")
 
     bridge = SpackCIBridge()
     bridge.setup_ssh(ssh_key_base64)
